+++ crapssim/strategy/__init__.py
@@ -6,7 +6,7 @@ after the previous bets and table have been updated.
 
 from crapssim.strategy.core import Strategy, AggregateStrategy, BetIfTrue, RemoveIfTrue, \
     IfBetNotExist, BetPointOff, BetPointOn, CountStrategy
-from crapssim.strategy.odds import OddsMultiplierStrategy
-from crapssim.strategy.examples import BetPassLine, PassLineOddsMultiplier, BetDontPass, \
-    DontPassOddsMultiplier, \
+from crapssim.strategy.odds import OddsMultiplierStrategy, PassLineOddsMultiplier, \
+    DontPassOddsMultiplier
+from crapssim.strategy.examples import BetPassLine, BetDontPass, \
     BetPlace
